Add the bin file's directory to the load path
As of ruby 1.9.2, the current directory of the invoked ruby file is no
longer added to the load path. When we extracted classes from
bin/git-pair into the lib directory, users of newer rubies started to
see an error when running git-pair. The same behavior also affects
git-pair-commit. This fix addresses the problem.